Decision Tree: Multiple fixes - pruning, tree_depth, viz
Commit includes following changes:
- Pruning is not performed when cp = 0 (default behavior)
- A particular bug is fixed: User input of max_depth starts from 0 and
the internal tree_depth starts from 1. This discrepancy was not taken into
account when tree train termination was checked leading to trees
containing only two leaf nodes on last level.
- Integer categorical variable is treated as ordered and hence is not
re-ordered. If the original ordering method is desired, then integer
column needs to be cast to TEXT.
- Visualization is improved: nodes with categorical feature splits only
provide the last value in the split, instead of the complete list.
This is consistent with the visualization in scikit-learn.

Closes #111