DT: Include rows with NULL features in training
JIRA: MADLIB-1095

This commit enables decision tree to include rows with NULL feature
values in the training dataset. Features that have NULL values are not
used during the training, but the features with non-null values are
used.

Note: Training of a level requires each row to go through the tree above
the level. If a row contains NULL value for a feature used for splitting
a node in the tree, the path for that row (either left or right) will
be determined by
1. using a surrogate feature (if surrogates are enabled) or
2. using the branch that had majority of rows assigned to it.

Closes #125