Bugfix: Fix multiple bugs and perf issue in grouping
JIRA: MADLIB-1100, MADLIB-1107

The default threshold valud is set to 1e-5 instead of a value
dependent on the number of vertices (1/num_vertices*1000). The
older docs said it would be (1/number_vertices*100), but latest
experiments showed even that would converge in about 5 iterations.
So using 1000 instead of 100 now.

Remove hard coded schema name in install check.

This commit includes changes for performance improvement too.
- It distributes intermediate and temp edge table by
both grouping cols and the dest vertex, hoping it would perform
better with MPP databases such as Greenplum and HAWQ.
- The query that inserts random_prob for nodes that have no
incoming edges was inefficient, especially with grouping.
Updated that query which significantly reduces the time with
grouping. That query is now run outside the for loop since the
graph will not change.

Closes #137