DT: Include NULL rows in count for termination check
An optimization was added to check after training a node
if its children would split further by analyzing the tuples going to
those nodes. There were two issues with this optimization:

1. To decide if a node won't split, it was checking if a node is too
pure (i.e. all responses are similar) *and* the number of tuples in the
node was less than a threshold (min_split). The check, however, should
be *or* since either one of those conditions should stop a node from
splitting. This commit corrects the logic.

2. When the primary split feature for a node is computed, the statistics
of rows going to the true and false side don't include the rows that have
NULL value for this split feature. These "NULL" rows can only be
included in the statistics during the next pass when the child node is
trained. This commit ensures that in the presence of NULL rows, this
optimization is disabled so that we don't terminate prematurely
after comparing with a lower count.

Closes #142