Control: Update 'optimizer' GUC only if editable
JIRA: MADLIB-1109

This is applicable only for the Greenplum platforms:

We disable/enable ORCA using the 'optimizer' GUC in some functions for
performance reasons. Greenplum has another GUC 'optimizer_control' which
allows the user to disable updates to the 'optimizer' GUC. Updating
'optimizer' when 'optimizer_control = off' leads to an ugly error.

This commit adds a check for the value of 'optimizer_control' and
updates 'optimizer' only if 'optimizer_control = on'. If the
'optimizer_control' GUC is not available then 'optimizer' is assumed to
be accessible.

Other changes:
- Removed disabling of optimizer in K-means and composite controller since
platform issue has been fixed.
- EnableOptimizer and EnableHashagg have been aptly renamed to
OptimizerControl and HashaggControl.
- Some unused debug time() function calls have been commented out.

Closes #157