DT: Assert stats index only when it's used
The stats_i >= 0 assert is true only when the feature_index is valid
(i.e. the node is not terminated). The assert fails when checked before
the feature_indices check, hence moving it within the if statement.
This is also recommended since stats_i is only used within that 'if'
block.