Refactor unittest for balanced_dataset module with proper mocking.
This commit removes the unit test code from the balanced_sample
production code which is now it's own python file.

Also the unittest file now properly mocks all the modules in
balanced_sample.py and can be run on it's own.

For now, I have created the unittest file in the same location as the
production file but we can move it.